+++ src/utils/getPreviousDemoFile.ts
@@ -1,0 +1,29 @@
+import { DemoFiles, Demos } from "../models";
+import { FileProvider } from "../services/FileProvider";
+
+export const getPreviousDemoFile = async (demoFile?: {
+  filePath: string;
+}): Promise<{
+  filePath: string;
+  demo: Demos;
+} | undefined> => {
+  if (!demoFile) {
+    return;
+  }
+
+  // Get the next demo file
+  const demoFiles: DemoFiles = await FileProvider.getFiles() || {};
+  const files = Object.keys(demoFiles).sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
+  const fileIdx = files.findIndex((file) => file === demoFile.filePath);
+  const previousIdx = fileIdx - 1;
+
+  if (previousIdx < 0) {
+    return;
+  }
+
+  const nextFile = files[previousIdx];
+  return {
+    filePath: nextFile,
+    demo: demoFiles[nextFile],
+  };
+};